requirements.txt
fastapi==0.115.6
langchain==0.3.11
langchain_community==0.3.11
langchain_huggingface==0.1.2
langserve==0.3.0
modelscope==1.17.1
ms_swift==2.3.0.post1
numpy==2.2.0
pyttsx3==2.90
Requests==2.32.3
sherpa_ncnn==2.1.10
sounddevice==0.5.0
torch==2.3.1+cu121
transformers==4.44.0
uvicorn==0.32.1
